Removing setPublicCDNPath
I'm still not convinced about this, but having both publicPath and publicCDNPath
was weird. When the publicCDNPath was set, publicPath did nothing, except that
it was used as the prefix in the manifest file. Now, there is no prefix at all
in the manifest file, which might prove to be weird in its own way.