Always outputting runtime.js
This is a subjective decision we made, because we thought it would cause
less "confusing" moments (i.e. when 2 different entries would not share
the same modules).

Unfortunately, the single runtime chunk causes a strange bug with
mini-css-extract plugin where, under some conditions, the output CSS
uses the chunk filename instead of "filename".