Re-working dev-server and https detection
If you pass --https at the command line, that overrides your
devServer.https config. In webpack-dev-server v3, that wasn't
a problem. But in v4, the certificate config was moved under devServer.https,
which can now be an object. This meant that if you passed --https at
the command line but also set devServer.https = { ... }, it would
be "re-set" back to "true" and your config would be lost.

The fix is to not require the --https flag and look at it or the user's
config to determine if the dev-server is running in https.