Do not allow webpack-dev-server to find an open port
This is unfortunate, as this is a nice feature of webpack-dev-server where
if port 8080 is open, it will try 8081, etc.

However, because we need to build a manifest.json file, we need to know what
the port is for sure at Webpack config build time. Even going through the same
async process of using the "find-port" package isn't really doable... as it
would make the entire config generation async (this may be possible, but would
change a ton of things in Encore).